Integrate UTKFaceUtil within imagenet_utk.py
